+++ spec/rubocop/cop/rspec/repeated_subject_call_spec.rb
@@ -1,0 +1,105 @@
+# frozen_string_literal: true
+
+RSpec.describe RuboCop::Cop::RSpec::RepeatedSubjectCall do
+  it 'registers an offense for a singular block' do
+    expect_offense(<<-RUBY)
+      RSpec.describe Foo do
+        it do
+          subject
+          expect { subject }.to not_change { Foo.count }
+          ^^^^^^^^^^^^^^^^^^ Calls to subject are memoized, this block is misleading
+        end
+      end
+    RUBY
+  end
+
+  it 'registers an offense for repeated blocks' do
+    expect_offense(<<-RUBY)
+      RSpec.describe Foo do
+        it do
+          expect { subject }.to change { Foo.count }
+          expect { subject }.to not_change { Foo.count }
+          ^^^^^^^^^^^^^^^^^^ Calls to subject are memoized, this block is misleading
+        end
+      end
+    RUBY
+  end
+
+  it 'registers an offense for nested blocks' do
+    expect_offense(<<-RUBY)
+      RSpec.describe Foo do
+        it do
+          expect(subject.a).to eq(3)
+          nested_block do
+            expect { on_shard(:europe) { subject } }.to not_change { Foo.count }
+            ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ Calls to subject are memoized, this block is misleading
+          end
+        end
+      end
+    RUBY
+  end
+
+  it 'registers an offense for custom subjects' do
+    expect_offense(<<-RUBY)
+      RSpec.describe Foo do
+        subject(:bar) { do_something }
+
+        it do
+          bar
+          expect { bar }.to not_change { Foo.count }
+          ^^^^^^^^^^^^^^ Calls to subject are memoized, this block is misleading
+        end
+      end
+    RUBY
+  end
+
+  it 'registers no offenses for no block' do
+    expect_no_offenses(<<~RUBY)
+      RSpec.describe Foo do
+        it do
+          expect(subject.a).to eq(3)
+          expect(subject.b).to eq(4)
+        end
+      end
+    RUBY
+  end
+
+  it 'registers no offenses for block first' do
+    expect_no_offenses(<<~RUBY)
+      RSpec.describe Foo do
+        it do
+          expect { subject }.to change { Foo.count }
+          expect(subject.b).to eq(4)
+        end
+      end
+    RUBY
+  end
+
+  it 'registers no offenses for different subjects' do
+    expect_no_offenses(<<-RUBY)
+      RSpec.describe Foo do
+        subject { do_something_else }
+        subject(:bar) { do_something }
+
+        it do
+          expect { bar }.to not_change { Foo.count }
+          expect { subject }.to not_change { Foo.count }
+        end
+      end
+    RUBY
+  end
+
+  it 'registers no offenses for multiple no description it blocks' do
+    expect_no_offenses(<<-RUBY)
+      RSpec.describe Foo do
+        it do
+          expect { subject }.to change { Foo.count }
+        end
+
+        it do
+          expect(subject.b).to eq(4)
+        end
+      end
+    RUBY
+  end
+end